+++ CHANGELOG.md
@@ -6,30 +6,45 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+- [Fixed fetching account twice on login. Populate shard from account.](https://github.com/ElrondNetwork/dapp-core/pull/530)
+- [Updated WalletConnect provider to 2.1.0-beta.6 and stability fixes](https://github.com/ElrondNetwork/dapp-core/pull/529)
+- [Refetch account data on network change](https://github.com/ElrondNetwork/dapp-core/pull/526)
+
 ## [[2.2.4](https://github.com/ElrondNetwork/dapp-core/pull/525)] - 2022-12-28
+
 - [Fixed extension plain login](https://github.com/ElrondNetwork/dapp-core/pull/523)
+
 ## [[2.2.3](https://github.com/ElrondNetwork/dapp-core/pull/510)] - 2022-12-16
+
 - [Fixed login on nativeAuth block failed](https://github.com/ElrondNetwork/dapp-core/pull/516)
 - [Changed CSS to exclude `!important;` and added className to transaction info components](https://github.com/ElrondNetwork/dapp-core/pull/517)
 - [Added missing actions to TransactionActionsEnum `AxiosInterceptorContext` response](https://github.com/ElrondNetwork/dapp-core/pull/518)
+
 ## [[2.2.2](https://github.com/ElrondNetwork/dapp-core/pull/512)] - 2022-12-16
+
 - [Fixed loginHooks to have same response and changed `AxiosInterceptorContext` response](https://github.com/ElrondNetwork/dapp-core/pull/511)
 
-
 ## [[2.2.1](https://github.com/ElrondNetwork/dapp-core/pull/510)] - 2022-12-16
+
 - [Changed logging in logic to allow signing custom token](https://github.com/ElrondNetwork/dapp-core/pull/509)
 - [Security issue fixed by substituting `anchorme` with linkify.js and fixed `getScamFlag` logic](https://github.com/ElrondNetwork/dapp-core/pull/508)
 
 ## [[2.2.0](https://github.com/ElrondNetwork/dapp-core/pull/507)] - 2022-12-14
+
 - [Added Axios interceptor wrapper to be used with `nativeAuth`](https://github.com/ElrondNetwork/dapp-core/pull/506)
 - [Changed `signTransactions.ts` import to resolve Next.js support](https://github.com/ElrondNetwork/dapp-core/pull/505)
 - [Upgraded `@elrondnetwork/erdjs-web-wallet-provider` to vesion 2.1.2](https://github.com/ElrondNetwork/dapp-core/pull/503)
 - [Added `nativeAuth` and Storybook](https://github.com/ElrondNetwork/dapp-core/pull/504)
 
 ## [[2.1.20](https://github.com/ElrondNetwork/dapp-core/pull/500)] - 2022-12-07
+
 - [Allow compatible versions of peer deps](https://github.com/ElrondNetwork/dapp-core/pull/499)
+
 ## [[2.1.19](https://github.com/ElrondNetwork/dapp-core/pull/498)] - 2022-12-05
+
 - [Fixed web-wallet login on FireFox](https://github.com/ElrondNetwork/dapp-core/pull/497)
+
 ## [[2.1.18](https://github.com/ElrondNetwork/dapp-core/pull/495)] - 2022-12-01
 
 - [Updated WalletConnect provider to 2.1.0-beta.5](https://github.com/ElrondNetwork/dapp-core/pull/494)