+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [[2.2.5](https://github.com/ElrondNetwork/dapp-core/pull/531)] - 2023-01-05
+  
 
 - [Fixed fetching account twice on login. Populate shard from account.](https://github.com/ElrondNetwork/dapp-core/pull/530)
 - [Updated WalletConnect provider to 2.1.0-beta.6 and stability fixes](https://github.com/ElrondNetwork/dapp-core/pull/529)